cleanup stuff

+++ src/schnetpack_tools/environment.py
@@ -35,19 +35,19 @@ class OpenCLEnvironmentProvider(BaseEnvironmentProvider):
 			self.initialize()
 
 		n_atoms = np.uint32(len(atoms))
-		cell = np.array(atoms.cell, order='C')
-		positions = np.array(atoms.positions, order='C')
+		cell = np.array(atoms.cell, dtype=np.float64, order='C')
+		positions = np.array(atoms.positions, dtype=np.float64, order='C')
 		scaled_positions = np.zeros((n_atoms,4), dtype=np.float64, order='C')
-		scaled_positions[:,0:3] = np.linalg.solve(cell.T, positions.T).T.astype(np.float64)
+		scaled_positions[:,0:3] = np.linalg.solve(cell.T, positions.T).T
 
 		neighborhood_idx = -np.ones((n_atoms,self.max_nbh), dtype=np.int32, order='C')
-		offsets = np.zeros((n_atoms,self.max_nbh), dtype=cl.cltypes.char3, order='C') # Change to integer to improve perf.
+		offsets = np.zeros((n_atoms,self.max_nbh), dtype=cl.cltypes.char3, order='C')
 
 		scaled_positions_buffer = cl.Buffer(self.ctx, self.mf.READ_ONLY  | self.mf.COPY_HOST_PTR, hostbuf=scaled_positions)
 		neighborhood_idx_buffer = cl.Buffer(self.ctx, self.mf.READ_WRITE | self.mf.COPY_HOST_PTR, hostbuf=neighborhood_idx)
-		offsets_buffer           = cl.Buffer(self.ctx, self.mf.READ_WRITE | self.mf.COPY_HOST_PTR, hostbuf=offsets)
+		offsets_buffer          = cl.Buffer(self.ctx, self.mf.READ_WRITE | self.mf.COPY_HOST_PTR, hostbuf=offsets)
 
-		cell = cell.astype(np.float64)
+		#cell = cell.astype(np.float64)
 		self.kernel.neighbor_list(self.queue, (n_atoms,), None, scaled_positions_buffer, offsets_buffer, neighborhood_idx_buffer,
 			n_atoms, self.max_nbh, self.cutoff,
 			cell[0,0], cell[0,1], cell[0,2],
@@ -64,7 +64,7 @@ class OpenCLEnvironmentProvider(BaseEnvironmentProvider):
 		if I == 0:
 			I = 1
 
-		offset = np.zeros((n_atoms,I,3), dtype=np.float64, order='C')
+		offset = np.zeros((n_atoms,I,3), dtype=np.int32, order='C')
 		e1.wait()
 		offset[:,:,0] = offsets[:,0:I]['x']
 		offset[:,:,1] = offsets[:,0:I]['y']
